Moved some user-facing macro definitions into declaration file. Renamed the files so all examples compile without any change
